pyproject.toml
[project]
name = "bruki"
version = "0.2"
description = "Analyze screenshot, saved-image, and camera-photo activity patterns."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "flask>=3.0",
  "matplotlib>=3.8",
  "numpy>=2.2",
  "pandas>=2.2",
  "pillow>=10.0",
  "pytesseract>=0.3.10",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "seaborn>=0.13",
]

[project.optional-dependencies]
ml = [
  "scikit-learn>=1.4",
  "torch>=2.2",
  "transformers>=5.1.0",
]
notebook = [
  "ipykernel>=7.2.0",
  "ipywidgets>=8.1.8",
  "umap-learn>=0.5.11",
  "hdbscan>=0.8.41",
]
social = [
  "cairosvg>=2.8",
]

[dependency-groups]
dev = [
  "nbstripout>=0.9.0",
  "pre-commit>=3.8",
  "ruff>=0.8.0",
  "ty>=0.0.17",
]

[project.scripts]
activity = "bruki.activity:main"
www = "bruki.server.api:main"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bruki"]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]

[tool.ty.src]
include = ["bruki"]
exclude = ["docs/social", "data"]
respect-ignore-files = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E221"]
